data_epic_capstone/pyproject.toml
[project]
name = "ai-agent-directory"
version = "0.1.0"
description = ""
authors = [
    {name = "team_data_engineering"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.33.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "black[jupyter] (>=25.1.0,<26.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "boto3 (>=1.38.27,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

